# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "User API",
    "version": "1.0.0",
    "description": "API documentation for user-related operations"
  },
  "servers": [
    {
      "url": "https://backend-production-7b98.up.railway.app"
    }
  ],
...
  "paths": {
...
    "/library/{libId}/getBook/{bookId}/id/{userId}": {
      "get": {
        "tags": ["Books"],
        "summary": "Retrieve a book by ID",
        "description": "Retrieve a book by its unique ID from the library.",
        "parameters": [
          {
            "in": "path",
            "name": "libId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The unique ID of the library"
          },
          {
            "in": "path",
            "name": "bookId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The unique ID of the book"
          },
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The unique ID of the user"
          }
        ],
        "responses": {
          "200": {
            "description": "A book object",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "bookId": {
                      "type": "string"
                    },
                    "pagesRead": {
                      "type": "number"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Book not found"
          },
          "500": {
            "description": "Cannot find the book, try again"
          }
        }
      }
    },
    "/library/{libId}/id/{userId}/getBook/{bookId}": {
      "get": {
        "tags": ["Books"],
        "summary": "Check if a book exists in the library",
        "description": "Check if a book exists in the library by its unique ID.",
        "parameters": [
          {
            "in": "path",
            "name": "libId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The unique ID of the library"
          },
          {
            "in": "path",
            "name": "bookId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The unique ID of the book"
          },
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The unique ID of the user"
          }
        ],
        "responses": {
          "200": {
            "description": "Book exists",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "result": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Book not found"
          },
          "500": {
            "description": "Cannot find the book, try again"
          }
        }
      }
    },
    "/getSpecLibrary/{libId}/id/{userId}": {
      "get": {
        "tags": ["Books"],
        "summary": "Retrieve a specific library by ID",
        "description": "Retrieve a specific library by its unique ID.",
        "parameters": [
          {
            "in": "path",
            "name": "libId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The unique ID of the library"
          },
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The unique ID of the user"
          }
        ],
        "responses": {
          "200": {
            "description": "A library object",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "libId": {
                      "type": "string"
                    },
                    "libName": {
                      "type": "string"
                    },
                    "books": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "bookId": {
                            "type": "string"
                          },
                          "pagesRead": {
                            "type": "number"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Library not found"
          },
          "500": {
            "description": "Cannot find the library, try again"
          }
        }
      }
    },
    "/getLibraries/{userId}": {
      "get": {
        "tags": ["Books"],
        "summary": "Retrieve all libraries for a user",
        "description": "Retrieve all libraries for a user by their unique ID.",
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The unique ID of the user"
          }
        ],
        "responses": {
          "200": {
            "description": "A list of libraries",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "libId": {
                        "type": "string"
                      },
                      "libName": {
                        "type": "string"
                      },
                      "books": {
                        "type": "array",
                        "items": {
                          "type": "object",
                          "properties": {
                            "bookId": {
                              "type": "string"
                            },
                            "pagesRead": {
                              "type": "number"
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "User not found"
          },
          "500": {
            "description": "Cannot find the libraries, try again"
          }
        }
      }
    },
    "/modifyPages": {
      "put": {
        "tags": ["Books"],
        "summary": "Modify the number of pages read",
        "description": "Modify the number of pages read for a specific book in a library.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "bookId": {
                    "type": "string"
                  },
                  "libId": {
                    "type": "string"
                  },
                  "pages": {
                    "type": "number"
                  },
                  "userId": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Number of pages read updated"
          },
          "404": {
            "description": "User or library not found"
          },
          "500": {
            "description": "Cannot modify the number of pages"
          }
        }
      }
    },
    "/createLibrary": {
      "post": {
        "tags": ["Books"],
        "summary": "Create a new library",
        "description": "Create a new library for a user.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "libName": {
                    "type": "string"
                  },
                  "libId": {
                    "type": "string"
                  },
                  "userId": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Library created successfully"
          },
          "404": {
            "description": "User not found"
          },
          "500": {
            "description": "Failed to create the library, try again"
          }
        }
      }
    },
    "/addBook": {
      "post": {
        "tags": ["Books"],
        "summary": "Add a book to a library",
        "description": "Add a new book to a specific library.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "bookId": {
                    "type": "string"
                  },
                  "libId": {
                    "type": "string"
                  },
                  "userId": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Book added successfully"
          },
          "404": {
            "description": "User or library not found"
          },
          "500": {
            "description": "Cannot add the book"
          }
        }
      }
    },
    "/library/{libId}/deleteBook/{bookId}/id/{userId}": {
      "delete": {
        "tags": ["Books"],
        "summary": "Delete a book from a library",
        "description": "Delete a book from a specific library by its unique ID.",
        "parameters": [
          {
            "in": "path",
            "name": "libId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The unique ID of the library"
          },
          {
            "in": "path",
            "name": "bookId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The unique ID of the book"
          },
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The unique ID of the user"
          }
        ],
        "responses": {
          "200": {
            "description": "Book deleted successfully"
          },
          "404": {
            "description": "User or library not found"
          },
          "500": {
            "description": "Cannot delete the book"
          }
        }
      }
    },
    "/deleteLibrary/{libId}/id/{userId}": {
      "delete": {
        "tags": ["Books"],
        "summary": "Delete a library",
        "description": "Delete a specific library by its unique ID.",
        "parameters": [
          {
            "in": "path",
            "name": "libId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The unique ID of the library"
          },
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The unique ID of the user"
          }
        ],
        "responses": {
          "200": {
            "description": "Library deleted successfully"
          },
          "404": {
            "description": "User or library not found"
          },
          "500": {
            "description": "Cannot delete the library"
          }
        }
      }
    }
  }
}